fix promise in response meta
