+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <a name="0.12.0"></a>
 # [0.12.0](https://github.com/ice-services/moleculer/compare/v0.11.10...v0.12.0) (2018-xx-xx)
 
-This version contains the most changes in the history of Moleculer! ~200 commits and a lot of features.
+This version contains the most changes in the history of Moleculer! More than 200 commits with 17k additions and a lot of new features.
 
 # Breaking changes
